# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me/credential": {
      "get": {
        "description": "Returns the credential kind, its granted scopes and what those scopes actually amount to once the owner’s roles are applied. The MCP server filters its tool list by `effectiveScopes`.",
        "operationId": "current",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentCredentialResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Describe the calling credential and its effective permissions",
        "tags": ["API tokens"]
      }
    },
...
    "/api/v1/tokens/permissions": {
      "get": {
        "description": "Drives the permission picker on the token-creation screen and the CLI’s --permissions validation.",
        "operationId": "permissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionCatalogResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the permission catalog and what the caller may grant",
        "tags": ["API tokens"]
      }
    },
...
    "/api/v1/roles": {
      "get": {
        "operationId": "findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by role name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all roles",
        "tags": ["Roles"]
      },
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "ROLE_002 — A role with this name already exists",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create role",
        "tags": ["Roles"]
      }
    },
    "/api/v1/roles/{id}": {
      "get": {
        "operationId": "findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "ROLE_001 — Role not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get role by ID",
        "tags": ["Roles"]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "ROLE_001 — Role not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a role (description and/or permissions)",
        "tags": ["Roles"]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "ROLE_003 — System roles cannot be deleted",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "ROLE_001 — Role not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete role",
        "tags": ["Roles"]
      }
    },
    "/api/v1/roles/{id}/permissions": {
      "put": {
        "operationId": "updatePermissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolePermissionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "ROLE_001 — Role not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace a role's permission set",
        "tags": ["Roles"]
      }
    },
...
    "/api/v1/leads": {
      "get": {
        "operationId": "list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeadResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the leads the caller can reach",
        "tags": ["Leads"]
      }
    },
    "/api/v1/leads/{id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "LEAD_001 — Lead not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one lead",
        "tags": ["Leads"]
      }
    },
...
  },
  "info": {
    "title": "Flama API",
    "description": "Flama REST API documentation",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InvalidParamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "email",
            "description": "Dotted path to the offending field"
          },
          "reason": {
            "type": "string",
            "example": "Invalid email",
            "description": "Why the value was rejected"
          }
        },
        "required": ["name", "reason"]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://flama.dev/errors#user_001",
            "description": "URI reference identifying the problem type. `about:blank` when the status code says everything."
          },
          "title": {
            "type": "string",
            "example": "User not found",
            "description": "Short summary, stable per problem type"
          },
          "status": {
            "type": "number",
            "example": 404,
            "description": "HTTP status code"
          },
          "detail": {
            "type": "string",
            "example": "No user with id 3f1c…",
            "description": "Explanation specific to this occurrence"
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/users/3f1c",
            "description": "URI reference of the specific occurrence — the request path"
          },
          "code": {
            "type": "string",
            "example": "USER_001",
            "description": "Machine-readable catalog code"
          },
          "correlationId": {
            "type": "string",
            "description": "Correlation id of the request, for log lookups"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-01-01T00:00:00.000Z",
            "description": "When it happened"
          },
          "invalidParams": {
            "description": "Per-field validation failures",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvalidParamDto"
            }
          }
        },
        "required": ["type", "title", "status"]
      },
...
      "CurrentCredentialResponseDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "description": "How the caller authenticated.",
            "enum": ["session", "api-token", "oauth"]
          },
          "userId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "grantedScopes": {
            "type": "array",
            "description": "Scopes the credential carries. Null for a browser session, which is not scope-restricted.",
            "nullable": true,
            "items": {
              "type": "string",
              "enum": [
                "profile:read",
                "profile:write",
                "users:read",
                "users:write",
                "admin:read",
                "admin:write",
                "roles:read",
                "roles:write",
                "organizations:read",
                "organizations:write",
                "members:read",
                "members:write",
                "invitations:read",
                "invitations:write",
                "workspaces:read",
                "workspaces:write",
                "tokens:read",
                "tokens:write",
                "billing:read",
                "billing:write",
                "leads:read",
                "leads:write"
              ]
            }
          },
          "effectiveScopes": {
            "type": "array",
            "description": "What the credential can actually do: its scopes intersected with the owner’s current roles.",
            "items": {
              "type": "string",
              "enum": [
                "profile:read",
                "profile:write",
                "users:read",
                "users:write",
                "admin:read",
                "admin:write",
                "roles:read",
                "roles:write",
                "organizations:read",
                "organizations:write",
                "members:read",
                "members:write",
                "invitations:read",
                "invitations:write",
                "workspaces:read",
                "workspaces:write",
                "tokens:read",
                "tokens:write",
                "billing:read",
                "billing:write",
                "leads:read",
                "leads:write"
              ]
            }
          },
          "organizationIds": {
            "description": "Organizations the credential is restricted to, or null when unrestricted.",
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "kind",
          "userId",
          "email",
          "grantedScopes",
          "effectiveScopes",
          "organizationIds",
          "expiresAt"
        ]
      },
...
      "PermissionCatalogResponseDto": {
        "type": "object",
        "properties": {
          "groups": {
            "description": "Every permission group, each with a Read and an Edit level. Render these as the permission picker.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionGroupDto"
            }
          },
          "grantable": {
            "type": "array",
            "description": "Scopes the caller may put on a token. Anything outside this list is refused at creation.",
            "example": ["profile:read", "users:read"],
            "items": {
              "type": "string",
              "enum": [
                "profile:read",
                "profile:write",
                "users:read",
                "users:write",
                "admin:read",
                "admin:write",
                "roles:read",
                "roles:write",
                "organizations:read",
                "organizations:write",
                "members:read",
                "members:write",
                "invitations:read",
                "invitations:write",
                "workspaces:read",
                "workspaces:write",
                "tokens:read",
                "tokens:write",
                "billing:read",
                "billing:write",
                "leads:read",
                "leads:write"
              ]
            }
          }
        },
        "required": ["groups", "grantable"]
      },
...
      "CreateRoleRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50,
            "pattern": "^[a-z0-9-_]+$"
          },
          "description": {
            "type": "string",
            "maxLength": 255
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "action": {
                  "type": "string",
                  "minLength": 1
                },
                "subject": {
                  "type": "string",
                  "minLength": 1
                },
                "conditions": {
                  "type": "object",
                  "additionalProperties": {}
                },
                "fields": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "inverted": {
                  "type": "boolean"
                },
                "reason": {
                  "type": "string"
                }
              },
              "required": ["action", "subject"]
            },
            "default": []
          }
        },
        "required": ["name"]
      },
      "RoleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isSystem": {
            "type": "boolean",
            "description": "System roles cannot be deleted or renamed."
          },
          "organizationId": {
            "type": "string",
            "description": "Owning organization, or null for a global role template shared by every tenant.",
            "nullable": true
          },
          "permissions": {
            "type": "array",
            "description": "CASL permission rules granted by this role.",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "isSystem",
          "permissions",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateRoleRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 255
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "action": {
                  "type": "string",
                  "minLength": 1
                },
                "subject": {
                  "type": "string",
                  "minLength": 1
                },
                "conditions": {
                  "type": "object",
                  "additionalProperties": {}
                },
                "fields": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "inverted": {
                  "type": "boolean"
                },
                "reason": {
                  "type": "string"
                }
              },
              "required": ["action", "subject"]
            }
          }
        }
      },
      "UpdateRolePermissionsRequest": {
        "type": "object",
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "action": {
                  "type": "string",
                  "minLength": 1
                },
                "subject": {
                  "type": "string",
                  "minLength": 1
                },
                "conditions": {
                  "type": "object",
                  "additionalProperties": {}
                },
                "fields": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "inverted": {
                  "type": "boolean"
                },
                "reason": {
                  "type": "string"
                }
              },
              "required": ["action", "subject"]
            }
          }
        },
        "required": ["permissions"]
      },
...
    }
  }
}